Version 1.1
1.1:
-	Added:
Nop, LangString, SetBrandingImage, the /a switch for File, the /e switch for SubSection,
the /o switch for Section, $PLUGINSDIR, !ifmacrodef, !ifmacrondef, IDD_LICENSE_FSRB, IDD_LICENSE_FSCB,
the /FINAL switch for SetCompressor and the /SD switch for MessageBox.
-	Changed:
InitPluginsDir, IfSilent, CreateFont and SetCtlColors are not attributes, they're commands.
EnabledBitmap and DisabledBitmap are obselete.
LoadLanguageFile is an instruction, it's... hmm... atrribute?
Uninstaller now has un.onMouseOverSection and un.onSelChange too.
There is no un.onInstSuccess and un.onInstFailed, it's un.onUninstSuccess and un.onUninstFailed.